requirements.txt
langchain
faiss-cpu
pandas
python-dotenv
streamlit
sentence-transformers
requests